Refactor Memsic MX2125 sensor to individual module.
