+++ src/modules/blockly/generators/propc/sensors.js
@@ -194,9 +194,11 @@ Blockly.Blocks.joystick_input_yaxis = {
     }
 
     this.chan = ['x', 'X'];
+
     if (this.type === 'joystick_input_yaxis') {
       this.chan = ['y', 'Y'];
     }
+
     this.setTooltip(Blockly.MSG_JOYSTICK_INPUT_TOOLTIP);
     this.setColour(colorPalette.getColor('input'));
     this.appendDummyInput()
@@ -226,6 +228,7 @@ Blockly.propc.joystick_input_yaxis = function() {
 
 Blockly.Blocks.joystick_input_xaxis = Blockly.Blocks.joystick_input_yaxis;
 Blockly.propc.joystick_input_xaxis = Blockly.propc.joystick_input_yaxis;
+
 
 // ---------------- Sound Impact Sensor Blocks -----------------------
 
@@ -1011,135 +1014,7 @@ Blockly.propc.fp_scanner_scan = function() {
 };
 
 
-// -------------Memsic Tilt/Accel (MX2125 Module) -------------------
-
-/**
- *
- * @type {{
- *  init: Blockly.Blocks.MX2125_acceleration_xaxis.init,
- *  addPinMenu: *,
- *  mutationToDom: *,
- *  helpUrl: string, setToOther: *,
- *  domToMutation: *
- *  }}
- */
-Blockly.Blocks.MX2125_acceleration_xaxis = {
-  helpUrl: Blockly.MSG_MEMSIC_HELPURL,
-  init: function() {
-    this.setTooltip(Blockly.MSG_MX2125_ACCELERATION_XAXIS_TOOLTIP);
-    this.setColour(colorPalette.getColor('input'));
-    this.chan = ['x', 'X'];
-    this.pinChoices = ['PIN'];
-    this.otherPin = [false];
-    if (this.type.indexOf('yaxis') > -1) {
-      this.chan = ['y', 'Y'];
-    }
-    this.pinChoices = ['PIN' + this.chan[1]];
-    this.res = ['acceleration', 'accel'];
-    if (this.type.indexOf('tilt') > -1) {
-      this.res = ['tilt', 'tilt'];
-    }
-    this.addPinMenu('Memsic ' + this.res[0] + ' ' + this.chan[0] +
-        '-axis PIN', null, 0);
-    this.setInputsInline(true);
-    this.setNextStatement(false, null);
-    this.setPreviousStatement(false, null);
-    this.setOutput(true, 'Number');
-  },
-  mutationToDom: Blockly.Blocks['sensor_ping'].mutationToDom,
-  domToMutation: Blockly.Blocks['sensor_ping'].domToMutation,
-  addPinMenu: Blockly.Blocks['sensor_ping'].addPinMenu,
-  setToOther: Blockly.Blocks['sensor_ping'].setToOther,
-};
-
-/**
- *
- * @return {[string, number]}
- * @constructor
- */
-Blockly.propc.MX2125_acceleration_xaxis = function() {
-  if (!this.disabled) {
-    Blockly.propc.definitions_['include_mx2125'] = '#include "mx2125.h"';
-  }
-  let pin = '0';
-  if (this.otherPin[0]) {
-    pin = Blockly.propc.valueToCode(
-        this, 'PIN' + this.chan[1], Blockly.propc.ORDER_ATOMIC) || '0';
-  } else {
-    pin = this.getFieldValue('PIN' + this.chan[1]);
-  }
-  return ['mx_' + this.res[1] + '(' + pin + ')', Blockly.propc.ORDER_NONE];
-};
-
-/**
- * Alias MX2125_acceleration_yaxis block to the MX2125_acceleration_xaxis block
- * @type {{
- *    init: Blockly.Blocks.MX2125_acceleration_xaxis.init,
- *    addPinMenu: *,
- *    mutationToDom: *,
- *    helpUrl: string,
- *    setToOther: *,
- *    domToMutation: *
- *  }}
- */
-Blockly.Blocks.MX2125_acceleration_yaxis =
-    Blockly.Blocks.MX2125_acceleration_xaxis;
-
-/**
- * Alias the MX2125_acceleration_yaxis C code generator to the
- * MX2125_acceleration_xaxis object
- * @type {function(): (string|number)[]}
- */
-Blockly.propc.MX2125_acceleration_yaxis =
-    Blockly.propc.MX2125_acceleration_xaxis;
-
-Blockly.Blocks.MX2125_tilt_xaxis = Blockly.Blocks.MX2125_acceleration_xaxis;
-Blockly.propc.MX2125_tilt_xaxis = Blockly.propc.MX2125_acceleration_xaxis;
-Blockly.Blocks.MX2125_tilt_yaxis = Blockly.Blocks.MX2125_acceleration_xaxis;
-Blockly.propc.MX2125_tilt_yaxis = Blockly.propc.MX2125_acceleration_xaxis;
-
-Blockly.Blocks.MX2125_rotation = {
-  helpUrl: Blockly.MSG_MEMSIC_HELPURL,
-  init: function() {
-    this.setTooltip(Blockly.MSG_MX2125_ROTATION_TOOLTIP);
-    this.setColour(colorPalette.getColor('input'));
-    this.pinChoices = ['PINX', 'PINY'];
-    this.otherPin = [false, false];
-    this.addPinMenu('Memsic rotation x-axis PIN', 'YAXIS', 0);
-    this.addPinMenu('y-axis PIN', 'XAXIS', 1);
-    this.setInputsInline(true);
-    this.setNextStatement(false, null);
-    this.setPreviousStatement(false, null);
-    this.setOutput(true, 'Number');
-  },
-  mutationToDom: Blockly.Blocks['sensor_ping'].mutationToDom,
-  domToMutation: Blockly.Blocks['sensor_ping'].domToMutation,
-  addPinMenu: Blockly.Blocks['sensor_ping'].addPinMenu,
-  setToOther: Blockly.Blocks['sensor_ping'].setToOther,
-};
-
-/**
- *
- * @return {[string, number]}
- * @constructor
- */
-Blockly.propc.MX2125_rotation = function() {
-  if (!this.disabled) {
-    Blockly.propc.definitions_['include_mx2125'] = '#include "mx2125.h"';
-  }
-  const pinVal = ['0', '0'];
-  for (let i = 0; i < this.pinChoices.length; i++) {
-    if (this.otherPin[i]) {
-      pinVal[i] = Blockly.propc.valueToCode(
-          this, this.pinChoices[i], Blockly.propc.ORDER_ATOMIC) || '0';
-    } else {
-      pinVal[i] = this.getFieldValue(this.pinChoices[i]);
-    }
-  }
-  const code = 'mx_rotate(' + pinVal[0] + ', ' + pinVal[1] + ')';
-  return [code, Blockly.propc.ORDER_NONE];
-};
-
+//
 // --------------Accelerometer (MMA7455 Module) Blocks--------------------
 
 /**
